dynamodb-firestore/copy-data/requirements.txt
google-cloud-firestore
google-cloud-datastore
google-cloud-storage
boto3
smart-open
more-itertools